Code Review 2 • Analysis of the program. - The program takes user data and saves it to firebase, I don’t see my profile post data yet. I know we are still trying to grasp it. • Was the program available in UC Github on time? -Yes it was • Is the program documented/commented well enough for you to understand? - Yes, variables are clear on what is going on in the project • Does the program compile? - Yes from unit test and app. • Rationale behind your changes. - It is hard for me to change much as they are very organized. I did noticed that an interface was not in the dao folder. I moved IUniversityService to dao. Ran a test and everything seemed to work. I assumed they are using two JSON streams as if I just use one the program gets errors. • Links to three specific commits that you made to your own group's GitHub repository after Sprint 1's deadline, and before Sprint 2's deadline.
• Three technical concepts
- Firebase implementation
- Implementing buttons to the project\
MainViewModel sending and gaining information from firebase.
