+++ UniConnectApp/app/src/main/java/com/example/uniconnect/AppModule.kt
@@ -1,6 +1,6 @@
 package com.example.uniconnect
 
-import com.example.uniconnect.service.IUniversityService
+import com.example.uniconnect.dao.IUniversityService
 import com.example.uniconnect.service.UniversityService
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
